requirements.txt
typing-extensions
langchain-openai
langchain-experimental
langgraph
langchain-google-genai
langchain-community
dashscope
pandas
feedparser
requests
chromadb
setuptools
rich
questionary
typer
python-dotenv
